refactor(runtime): split host domain runtime module
