+++ apps/desktop/src-tauri/crates/host-domain/src/runtime/mod.rs
@@ -1,0 +1,22 @@
+mod events;
+mod registry;
+mod route;
+mod state;
+
+pub use events::{
+    DevServerEvent, DevServerGroupState, DevServerScriptState, DevServerScriptStatus,
+    DevServerTerminalChunk, RunEvent,
+};
+pub use registry::{
+    builtin_runtime_registry, default_runtime_kind, AgentRuntimeKind, RuntimeCapabilities,
+    RuntimeDefinition, RuntimeDescriptor, RuntimeProvisioningMode, RuntimeRegistry,
+    RuntimeStartupReadinessConfig, RuntimeSupportedScope,
+};
+pub use route::RuntimeRoute;
+pub use state::{
+    BuildContinuationTarget, BuildContinuationTargetSource, RepoRuntimeHealthCheck,
+    RepoRuntimeHealthMcp, RepoRuntimeHealthObservation, RepoRuntimeHealthRuntime,
+    RepoRuntimeHealthState, RepoRuntimeMcpStatus, RepoRuntimeStartupFailureKind,
+    RepoRuntimeStartupStage, RepoRuntimeStartupStatus, RunState, RunSummary,
+    RuntimeInstanceSummary, RuntimeRole,
+};
